IInteractable.GetInteractionSettings
Main reason behind this was player stopping when interacting with toilet door. While pressing movement keys and also pressing interaction key was triggering the disable process of CharacterMovement input. Also Interaction was disabled rather or not it is required to disable. This way we are able to define what to do while interacting with the IInteractable